+++ routers/index.js
@@ -11,6 +11,9 @@
   */
 "use strict";
 
+const {resolve} = require("path");
+const [user, workflow, fs] = [require(resolve(__dirname, "user")), require(resolve(__dirname, "workflow")), require(resolve(__dirname, "fileSystem"))];
+
 module.exports = _ => {
   _.get("/test", function* (next) {
     this.render("index", {
@@ -21,6 +24,16 @@ module.exports = _ => {
     return yield next;
   });
 
+  // 用户模块
+  _.get("/user/list", user.list); // 用户信息列表
+
+  // 工作流模块
+  _.get("/workflow/list", workflow.list); // 工作流信息列表
+
+  // 文件系统模块
+  _.get("/fs/list", fs.list); // 更新文件信息列表
+
+ // 测试：开发者信息
   _.get("/author", function* (next) {
     this.body = JSON.stringify({
         name: "afterloe",