+++ routers/workflow.js
@@ -1,0 +1,21 @@
+/**
+  * afterloe - cynomy_portal_server/routers/workflow.js
+  *
+  * Copyright(c) afterloe.
+  * MIT Licensed
+  *
+  * Authors:
+  *   afterloe <afterloeliu@jwis.cn> (https://github.com/afterloe)
+  * Date:
+  *   2017-2-2 16:14:38
+  */
+"use strict";
+
+const list = function* (next) {
+
+  return yield next;
+};
+
+module.exports = {
+  list,
+};