+++ History.md
@@ -901,3 +901,20 @@ commit by afterloe (lm6289511@gmail.com)
 > servers/websocket.js  
 > tools/_util.html  
 
+
+1.0.3 / 2017-2-2 18:24:50
+==================
+commit by afterloe (lm6289511@gmail.com)
+
+  * master: http获取接口调试
+
+
+### new files
+> routers/fileSystem.js  
+> routers/user.js  
+> routers/workflow.js  
+
+### modify files
+> routers/index.js  
+> servers/registry.js  
+